# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Account API",
        "description": "Dokumentasi API untuk manajemen akun",
        "version": "1.0.0"
    },
    "paths": {
        "/api/accounts": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Ambil semua akun",
                "operationId": "13efa5706d1fae14fc1cf87c9f6b1427",
                "responses": {
                    "200": {
                        "description": "Berhasil ambil semua akun"
                    }
                }
            },
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Tambah akun baru",
                "operationId": "e974666bc6c050296a8ee09a3ab6533f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "age"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "age": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User berhasil ditambahkan"
                    }
                }
            }
        },
        "/api/accounts/{id}": {
            "put": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Update akun",
                "operationId": "9be8085be375b01725706b78df130bca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "age"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "age": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User berhasil diupdate"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Hapus akun",
                "operationId": "102c60c5fd1e94be8c71c4efbeb51bf8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User berhasil dihapus"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Accounts",
            "description": "Endpoints untuk operasi akun"
        }
    ]
}